requirements.txt
pyserial>=3.5
gpio>=0.3.0